Fixes #2 fixes #8 @gholtslander
Added the script I used to generate the migration as well as the migration code itself. If you go to /question/migration it will bring you to a confirmation page screenshotted below:

Clicking confirm will create all 180 questions and put them in the database, then it will print 'Created the questions' and that'll be that.